requirements.txt
kafka-python==2.0.2
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
httpx==0.27.0
pydantic==2.6.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
